+++ packages/frontend/component/src/lit-react/index.ts
@@ -1,4 +1,7 @@
-export { createComponent as createReactComponentFromLit } from './create-component';
+export {
+  createComponent as createReactComponentFromLit,
+  type ReactWebComponent,
+} from './create-component';
 export * from './lit-portal';
 export { toReactNode } from './to-react-node';
 export { templateToString } from './utils';
