+++ packages/frontend/core/src/blocksuite/block-suite-editor/specs/edgeless.ts
@@ -1,23 +1,17 @@
 import { enableAIExtension } from '@affine/core/blocksuite/ai';
-import { FeatureFlagService } from '@affine/core/modules/feature-flag';
+import { enableAffineExtension } from '@affine/core/blocksuite/extensions';
 import { builtInTemplates as builtInEdgelessTemplates } from '@affine/templates/edgeless';
 import { builtInTemplates as builtInStickersTemplates } from '@affine/templates/stickers';
 import type { SpecBuilder, TemplateManager } from '@blocksuite/affine/blocks';
 import { EdgelessTemplatePanel, SpecProvider } from '@blocksuite/affine/blocks';
 import { type FrameworkProvider } from '@toeverything/infra';
 
-import { enableAffineExtension } from './custom/root-block';
-
 export function createEdgelessModeSpecs(
   framework: FrameworkProvider
 ): SpecBuilder {
-  const featureFlagService = framework.get(FeatureFlagService);
-  const enableAI = featureFlagService.flags.enable_ai.value;
   const edgelessSpec = SpecProvider._.getSpec('edgeless');
-  enableAffineExtension(framework, edgelessSpec);
-  if (enableAI) {
-    enableAIExtension(edgelessSpec, framework);
-  }
+  enableAffineExtension(edgelessSpec, framework);
+  enableAIExtension(edgelessSpec, framework);
 
   return edgelessSpec;
 }
