+++ packages/frontend/core/src/blocksuite/editors/index.ts
@@ -1,8 +1,27 @@
+import { createReactComponentFromLit } from '@affine/component';
+import { DocTitle } from '@blocksuite/affine/blocks';
+import React from 'react';
+
 import { EdgelessEditor } from './edgeless-editor';
 import { PageEditor } from './page-editor';
 
 export * from './edgeless-editor';
 export * from './page-editor';
+
+export const LitDocEditor = createReactComponentFromLit({
+  react: React,
+  elementClass: PageEditor,
+});
+
+export const LitDocTitle = createReactComponentFromLit({
+  react: React,
+  elementClass: DocTitle,
+});
+
+export const LitEdgelessEditor = createReactComponentFromLit({
+  react: React,
+  elementClass: EdgelessEditor,
+});
 
 export function effects() {
   customElements.define('page-editor', PageEditor);
