+++ packages/frontend/core/src/blocksuite/extensions/editor-config/database.ts
