+++ packages/frontend/core/src/blocksuite/extensions/editor-config/linked.ts
