+++ packages/frontend/core/src/blocksuite/extensions/editor-config/toolbar/copy-as-image.ts
