+++ packages/frontend/core/src/blocksuite/extensions/editor-config/toolbar/index.ts
