+++ packages/frontend/core/src/blocksuite/extensions/electron-clipboard.ts
@@ -1,0 +1,10 @@
+import { DesktopApiService } from '@affine/core/modules/desktop-api';
+import { NativeClipboardExtension } from '@blocksuite/affine/blocks';
+import type { FrameworkProvider } from '@toeverything/infra';
+
+export function patchForClipboardInElectron(framework: FrameworkProvider) {
+  const desktopApi = framework.get(DesktopApiService);
+  return NativeClipboardExtension({
+    copyAsPNG: desktopApi.handler.clipboard.copyAsPNG,
+  });
+}
