+++ packages/frontend/core/src/blocksuite/extensions/index.ts
@@ -1,0 +1,2 @@
+export * from './entry/enable-affine';
+export * from './entry/enable-mobile';
