+++ packages/frontend/core/src/blocksuite/extensions/keyboard-toolbar-config.ts
