+++ packages/frontend/core/src/blocksuite/extensions/note-config/edgeless-note-header.css.ts
