+++ packages/frontend/core/src/blocksuite/extensions/side-bar-service.ts
@@ -1,0 +1,20 @@
+import { WorkbenchService } from '@affine/core/modules/workbench';
+import { SidebarExtension } from '@blocksuite/affine/blocks';
+import type { FrameworkProvider } from '@toeverything/infra';
+
+export function patchSideBarService(framework: FrameworkProvider) {
+  const { workbench } = framework.get(WorkbenchService);
+
+  return SidebarExtension({
+    open: (tabId?: string) => {
+      workbench.openSidebar();
+      workbench.activeView$.value.activeSidebarTab(tabId ?? null);
+    },
+    close: () => {
+      workbench.closeSidebar();
+    },
+    getTabIds: () => {
+      return workbench.activeView$.value.sidebarTabs$.value.map(tab => tab.id);
+    },
+  });
+}
