+++ packages/frontend/core/src/blocksuite/extensions/theme.ts
@@ -1,0 +1,94 @@
+import { DocService, DocsService } from '@affine/core/modules/doc';
+import { AppThemeService } from '@affine/core/modules/theme';
+import { LifeCycleWatcher, StdIdentifier } from '@blocksuite/affine/block-std';
+import type { Signal, ThemeExtension } from '@blocksuite/affine/blocks';
+import {
+  ColorScheme,
+  createSignalFromObservable,
+  ThemeExtensionIdentifier,
+} from '@blocksuite/affine/blocks';
+import type { Container } from '@blocksuite/affine/global/di';
+import { type FrameworkProvider } from '@toeverything/infra';
+import type { Observable } from 'rxjs';
+import { combineLatest, map } from 'rxjs';
+
+export function getThemeExtension(
+  framework: FrameworkProvider
+): typeof LifeCycleWatcher {
+  class AffineThemeExtension
+    extends LifeCycleWatcher
+    implements ThemeExtension
+  {
+    static override readonly key = 'affine-theme';
+
+    private readonly themes: Map<string, Signal<ColorScheme>> = new Map();
+
+    protected readonly disposables: (() => void)[] = [];
+
+    static override setup(di: Container) {
+      super.setup(di);
+      di.override(ThemeExtensionIdentifier, AffineThemeExtension, [
+        StdIdentifier,
+      ]);
+    }
+
+    getAppTheme() {
+      const keyName = 'app-theme';
+      const cache = this.themes.get(keyName);
+      if (cache) return cache;
+
+      const theme$: Observable<ColorScheme> = framework
+        .get(AppThemeService)
+        .appTheme.theme$.map(theme => {
+          return theme === ColorScheme.Dark
+            ? ColorScheme.Dark
+            : ColorScheme.Light;
+        });
+      const { signal: themeSignal, cleanup } =
+        createSignalFromObservable<ColorScheme>(theme$, ColorScheme.Light);
+      this.disposables.push(cleanup);
+      this.themes.set(keyName, themeSignal);
+      return themeSignal;
+    }
+
+    getEdgelessTheme(docId?: string) {
+      const doc =
+        (docId && framework.get(DocsService).list.doc$(docId).getValue()) ||
+        framework.get(DocService).doc;
+
+      const cache = this.themes.get(doc.id);
+      if (cache) return cache;
+
+      const appTheme$ = framework.get(AppThemeService).appTheme.theme$;
+      const docTheme$ = doc.properties$.map(
+        props => props.edgelessColorTheme || 'system'
+      );
+      const theme$: Observable<ColorScheme> = combineLatest([
+        appTheme$,
+        docTheme$,
+      ]).pipe(
+        map(([appTheme, docTheme]) => {
+          const theme = docTheme === 'system' ? appTheme : docTheme;
+          return theme === ColorScheme.Dark
+            ? ColorScheme.Dark
+            : ColorScheme.Light;
+        })
+      );
+      const { signal: themeSignal, cleanup } =
+        createSignalFromObservable<ColorScheme>(theme$, ColorScheme.Light);
+      this.disposables.push(cleanup);
+      this.themes.set(doc.id, themeSignal);
+      return themeSignal;
+    }
+
+    override unmounted() {
+      this.dispose();
+    }
+
+    dispose() {
+      this.disposables.forEach(dispose => dispose());
+    }
+  }
+
+  return AffineThemeExtension;
+}
